fix damage

+++ src/main/java/dev/xkmc/l2enchantments/content/enchantments/sword/SwordEnchant.java
@@ -39,7 +39,7 @@ public class SwordEnchant extends BaseEnchantment {
 	public void onTargetAttacked(int lv, LivingAttackEvent event, AttackEventHandler.AttackCache attackCache) {
 	}
 
-	public int getAdditionalDamage(int lv, LivingHurtEvent event, AttackEventHandler.AttackCache attackCache) {
+	public double getAdditionalDamage(int lv, LivingHurtEvent event, AttackEventHandler.AttackCache attackCache) {
 		return 0;
 	}
